update(src/main.zig): unshare mount ns and make propagation private to enforce isolation
As described in https://lwn.net/Articles/689856/, when a new mount namespace is
created, it receives a copy of the mount points from the caller. By default, in
the initial mount namespace, everything is mounted with the MS_SHARED propagation
type. This means that every mount event will be reflected in each mount namespace
belonging to the same group. Since we created our mount namespace from the init
namespace, and we don't want mount events to be reflected on our filesystem view
to create isolation, we can recursively remount as private the root of our fs.
This is the difference that stands out also when issuing commands like:
sudo unshare -m
vs
sudo unshare -m --propagation unchanged (this calls a mount syscall like described
above)
and inspecting their syscalls.

Signed-off-by: Lorenzo Susini <susinilorenzo1@gmail.com>