update: clone into new pid namespace
Also move back the creation of the mount namespace to the clone syscall.
As per man page, the clone allows the child to directly be in the new pid namespace,
while when doing unshare, only its children will be in the new namespace.
For this reasons, it looks easier to have them all in the clone

Signed-off-by: Lorenzo Susini <susinilorenzo1@gmail.com>